requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tqdm>=4.62.0
tensorboard>=2.7.0
albumentations>=1.1.0
facenet-pytorch>=2.5.2
timm>=0.6.0
pyyaml>=6.0
pandas>=1.3.0
mtcnn>=0.1.1